move Transaction to primitives
